pyproject.toml
[project]
name = "jianzhi-ui-test"
version = "0.1.0"
description = "AI + Playwright UI 自动化测试 POC"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.1.0",
    "playwright>=1.40.0",
    "langchain-core>=0.3.0",
    "dashscope>=1.20.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
